requirements.txt

torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
pytest>=7.0.0
pydantic>=2.0.0
accelerate
bitsandbytes
scipy
groq
yfinance
lxml
matplotlib
seaborn
python-dotenv
